% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_information_helpers.R
\name{add_fivethirtyeight_predictions}
\alias{add_fivethirtyeight_predictions}
\title{Add FiveThirtyEight's Elo Predictions}
\usage{
add_fivethirtyeight_predictions(espnNFLscraper_dataset)
}
\arguments{
\item{espnNFLscraper_dataset}{: a dataset created by espnNFLscraper}
}
\value{
dataframe with 538's predictions appended.
}
\description{
Add the pregame predictions of FiveThirtyEight's Elo projection system
}
\examples{
get_game_details(espn_gameids = espn_nfl_ids \%>\% filter(season == 2019 & season_type == 2) \%>\% pull(espn_gameid)) \%>\% add_fivethirtyeight_predictions()
}
