% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_information_helpers.R
\name{add_gamescores}
\alias{add_gamescores}
\title{Add Gamescores}
\usage{
add_gamescores(espnNFLscraper_dataset)
}
\arguments{
\item{espnNFLscraper_dataset}{: a dataset created by espnNFLscraper}
}
\value{
dataframe with the outcomes of completed NFL games.
}
\description{
Add the outcomes of completed NFL games and discover if the home team won.
}
\examples{
get_game_details(espn_gameids = espn_nfl_ids \%>\% filter(season == 2019 & season_type == 2) \%>\% pull(espn_gameid)) \%>\% add_gamescores()
}
