% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_ids.R
\name{get_game_ids}
\alias{get_game_ids}
\title{Get Game IDs}
\usage{
get_game_ids(season, season_type = c("preseason", "regular", "postseason"))
}
\arguments{
\item{season:}{the specific season for which you want to grab gameids from ESPN}

\item{season_type:}{"preseason", "regular", or "postseason".}
}
\value{
dataframe with basic game info and espn gameids
}
\description{
Pull ESPN's GameIDs for NFL Games
}
\examples{
get_game_ids(season = 2019, season_type = "regular")
}
