% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_wp.R
\name{get_game_wp}
\alias{get_game_wp}
\title{Get WP from NFL Game}
\usage{
get_game_wp(espn_gameids)
}
\arguments{
\item{espn_gameids}{the gameids from espn}
}
\value{
dataframe with basic espn's WP for each team after plays in the game
}
\description{
Pull ESPN's WP at each point in the game. Function created by github.com/mrcaseb
}
\examples{
get_game_wp(espn_gameids = espn_nfl_ids \%>\% filter(season == 2019 & season_type == 2 & week == 1) \%>\% pull(espn_gameid))
}
