% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pregame_predictions.R
\name{get_pregame_predictions}
\alias{get_pregame_predictions}
\title{Get Pregame Predictions}
\usage{
get_pregame_predictions(espn_gameids)
}
\arguments{
\item{espn_gameids}{the gameids from espn}
}
\value{
dataframe with basic game info and pregame home team win probabilities
}
\description{
Pull and combine pregame predictions from multiple games
}
\examples{
get_pregame_predictions(espn_gameids = espn_nfl_ids \%>\% filter(season == 2019 & season_type == 2) \%>\% pull(espn_gameid))
}
